Fixed player existence through the game.
