+++ Assets/Scripts/Lobby&Elevator/Elevator.cs
@@ -1,6 +1,8 @@
 using System.Collections;
 using System.Collections.Generic;
+
 using UnityEngine;
+
 using ABOGGUS.Gameplay;
 
 public class Elevator : MonoBehaviour
