+++ Assets/Scripts/Menus/InventoryMenu.cs
@@ -1,8 +1,11 @@
 using System.Collections;
 using System.Collections.Generic;
+
 using UnityEngine;
-using ABOGGUS.PlayerObjects;
 using UnityEngine.UI;
+
+using ABOGGUS.PlayerObjects;
+using ABOGGUS.Gameplay;
 
 namespace ABOGGUS.Menus
 {
@@ -43,7 +46,7 @@ namespace ABOGGUS.Menus
         private void OpenInventory()
         {
             inventoryMenu.SetActive(true);
-            if (Player.key)
+            if (GameController.player.inventory.key)
             {
                 keyContainer.enabled = true;
             } else
