+++ Assets/Scripts/PlayerObjects/PlayerInventory.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 5312a72ad0e278a44ac88501b59e8dcf
+guid: 27bc5242700af194694db66c103ddba0
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
