+++ Assets/Scripts/PlayerObjects/States/PlayerFacingForward.cs
@@ -7,20 +7,20 @@ namespace ABOGGUS.PlayerObjects
     public class PlayerFacingForward : IPlayerState
     {
         private float yRot = 0;
-        Player player;
+        GameObject physicalGameObject;
 
-        public PlayerFacingForward(Player player)
+        public PlayerFacingForward(PlayerController playerController)
         {
             yRot = Rotator.cameraYRot;
-            this.player = player;
-            this.player.transform.localRotation = Quaternion.Euler(new Vector3(0, yRot, 0));
+            this.physicalGameObject = playerController.GetGameObject();
+            this.physicalGameObject.transform.localRotation = Quaternion.Euler(new Vector3(0, yRot, 0));
         }
         public void Move()
         {
             yRot = Rotator.cameraYRot;
-            player.transform.localRotation = Quaternion.Euler(new Vector3(0, yRot, 0));
-            Vector3 target = player.transform.position + player.transform.forward * PlayerController.speed;
-            player.transform.localPosition = Vector3.MoveTowards(player.transform.localPosition, target, PlayerController.speed);
+            physicalGameObject.transform.localRotation = Quaternion.Euler(new Vector3(0, yRot, 0));
+            Vector3 target = physicalGameObject.transform.position + physicalGameObject.transform.forward * PlayerController.speed;
+            physicalGameObject.transform.localPosition = Vector3.MoveTowards(physicalGameObject.transform.localPosition, target, PlayerController.speed);
         }
     }
 }
